roots: move to be a submodule of project
`roots` needs to be able to access private fields of project.
